+++ src/main/java/com/sysmap/srcmsportability/application/ports/in/PortabilityService.java
@@ -1,8 +1,7 @@
 package com.sysmap.srcmsportability.application.ports.in;
 
-import com.sysmap.srcmsportability.domain.entities.Portability;
-import com.sysmap.srcmsportability.domain.entities.User;
-import com.sysmap.srcmsportability.domain.entities.enums.StatusPortability;
+import com.sysmap.srcmsportability.application.ports.in.entities.Portability;
+import com.sysmap.srcmsportability.application.ports.in.entities.enums.StatusPortability;
 import org.springframework.data.crossstore.ChangeSetPersister;
 import org.springframework.http.ResponseEntity;
 
@@ -10,7 +9,7 @@ import java.util.UUID;
 
 public interface PortabilityService {
 
-    public Portability createPortability(Portability portability, User user);
+    public Portability createPortability(Portability portability);
 
     public ResponseEntity putStatusPortability(UUID portabilityId, StatusPortability status) throws ChangeSetPersister.NotFoundException;
