+++ src/main/java/com/sysmap/srcmsportability/application/ports/in/UserService.java
@@ -1,0 +1,9 @@
+package com.sysmap.srcmsportability.application.ports.in;
+
+import com.sysmap.srcmsportability.application.ports.in.entities.User;
+
+public interface UserService {
+
+    public User createUser(User user);
+
+}
