+++ src/main/java/com/sysmap/srcmsportability/application/ports/in/entities/LineInformation.java
@@ -1,0 +1,23 @@
+package com.sysmap.srcmsportability.application.ports.in.entities;
+
+import lombok.Data;
+import org.hibernate.annotations.GenericGenerator;
+
+import javax.persistence.*;
+import java.util.UUID;
+
+@Entity
+@Data
+@Table(name = "tb_line")
+public class LineInformation {
+
+    @Id
+    @GeneratedValue(generator = "UUID")
+    @GenericGenerator(
+            name = "UUID",
+            strategy = "org.hibernate.id.UUIDGenerator"
+    )
+    @Column(name = "line_id")
+    private UUID lineId;
+    private String number;
+}
