+++ src/main/java/com/sysmap/srcmsportability/application/ports/in/entities/enums/CellPhoneOperator.java
@@ -1,0 +1,6 @@
+package com.sysmap.srcmsportability.application.ports.in.entities.enums;
+
+public enum CellPhoneOperator {
+    VIVO,
+    CLARO
+}
