+++ src/main/java/com/sysmap/srcmsportability/application/service/UserServiceImpl.java
@@ -1,0 +1,20 @@
+package com.sysmap.srcmsportability.application.service;
+
+import com.sysmap.srcmsportability.application.ports.in.UserService;
+import com.sysmap.srcmsportability.application.ports.in.entities.User;
+import com.sysmap.srcmsportability.application.ports.out.UserRepository;
+
+public class UserServiceImpl implements UserService {
+
+    private final UserRepository userRepository;
+
+    public UserServiceImpl(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
+
+    @Override
+    public User createUser(User user) {
+        return this.userRepository.saveUser(user);
+    }
+
+}
