--- src/main/java/com/sysmap/srcmsportability/domain/entities/LineInformation.java
@@ -1,18 +1,0 @@
-package com.sysmap.srcmsportability.domain.entities;
-
-import lombok.Data;
-
-import javax.persistence.Entity;
-import javax.persistence.Id;
-import javax.persistence.Table;
-import java.util.UUID;
-
-@Entity
-@Data
-@Table(name = "tb_line")
-public class LineInformation {
-
-    @Id
-    private UUID lineId;
-    private String number;
-}
