--- src/main/java/com/sysmap/srcmsportability/domain/entities/enums/CellPhoneOperator.java
@@ -1,6 +1,0 @@
-package com.sysmap.srcmsportability.domain.entities.enums;
-
-public enum CellPhoneOperator {
-    VIVO,
-    CLARO
-}
