+++ src/main/java/com/sysmap/srcmsportability/framework/adapters/config/BeanConfiguration.java
@@ -2,7 +2,9 @@ package com.sysmap.srcmsportability.framework.adapters.config;
 
 import com.sysmap.srcmsportability.SrcMsPortabilityApplication;
 import com.sysmap.srcmsportability.application.ports.out.PortabilityRepository;
+import com.sysmap.srcmsportability.application.ports.out.UserRepository;
 import com.sysmap.srcmsportability.application.service.PortabilityServiceImpl;
+import com.sysmap.srcmsportability.application.service.UserServiceImpl;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
@@ -16,4 +18,8 @@ public class BeanConfiguration {
         return new PortabilityServiceImpl(portabilityRepository);
     }
 
+    @Bean
+    UserServiceImpl userServiceImpl(UserRepository userRepository) {
+        return new UserServiceImpl(userRepository);
+    }
 }
