+++ src/main/java/com/sysmap/srcmsportability/framework/adapters/in/dto/InputPortability.java
@@ -1,7 +1,9 @@
 package com.sysmap.srcmsportability.framework.adapters.in.dto;
 
-import com.sysmap.srcmsportability.domain.entities.Portability;
-import com.sysmap.srcmsportability.domain.entities.User;
+import com.sysmap.srcmsportability.application.ports.in.entities.Address;
+import com.sysmap.srcmsportability.application.ports.in.entities.LineInformation;
+import com.sysmap.srcmsportability.application.ports.in.entities.Portability;
+import com.sysmap.srcmsportability.application.ports.in.entities.User;
 import lombok.Data;
 
 @Data
@@ -9,4 +11,6 @@ public class InputPortability {
 
     private User user;
     private Portability portability;
+    private Address address;
+    private LineInformation lineInformation;
 }
