+++ src/main/java/com/sysmap/srcmsportability/framework/adapters/in/dto/InputPutStatus.java
@@ -1,6 +1,6 @@
 package com.sysmap.srcmsportability.framework.adapters.in.dto;
 
-import com.sysmap.srcmsportability.domain.entities.enums.StatusPortability;
+import com.sysmap.srcmsportability.application.ports.in.entities.enums.StatusPortability;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
